Added FavoritesEmpty component.
